--- a/bin/rake
+++ b/bin/rake
@@ -1,4 +1,5 @@
 #!/usr/bin/env ruby
+load File.expand_path("spring", __dir__) unless ENV['DISABLE_SPRING']
 require_relative '../config/boot'
 require 'rake'
 Rake.application.run