--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -7,6 +7,7 @@ Rails.application.configure do
   # note, newer versions of rails replace config.cache_classes with enable_reloading
   if (ENV['DISABLE_SPRING'].presence || '0') != '0'
     # spring enabled
+    raise 'spring should be disabled'
     config.cache_classes = false
   else
     config.cache_classes = true