--- a/config/spring.rb
+++ b/config/spring.rb
@@ -1,6 +1,0 @@
-Spring.watch(
-  ".ruby-version",
-  ".rbenv-vars",
-  "tmp/restart.txt",
-  "tmp/caching-dev.txt"
-)