--- a/spec/models/simple_spec.rb
+++ b/spec/models/simple_spec.rb
@@ -1,8 +1,0 @@
-
-require 'rails_helper'
-
-RSpec.describe 'hello world test', type: :model do
-  it "is okay" do
-    expect(1).to eq(1)
-  end
-end