--- a/spec/rails_helper.rb
+++ b/spec/rails_helper.rb
@@ -70,28 +70,36 @@ RSpec.configure do |config|
   config.include S3Utils
 
   config.before(:suite) do
-    Dir.glob('{drivers,spec}/**/fixpoints/*.yml').each do |filename|
-      FileUtils.rm(filename)
-    end
-    Dir.glob('{drivers,spec}/**/fixpoints/*.sql').each do |filename|
-      # The 2024 test kit is HUGE, so we'll use the preprocessed version
-      # This means when the test kit changes, we need to update both the CSVs and the SQL fixpoint
-      next if filename.include?('drivers/datalab_testkit/spec/fixpoints')
-
-      FileUtils.rm(filename)
-    end
-
-    GrdaWarehouse::Utility.clear!
-    Delayed::Job.delete_all
-    GrdaWarehouse::WarehouseReports::ReportDefinition.maintain_report_definitions
-    AccessGroup.maintain_system_groups
-
-    # disable papertrail for test performance
+    # disable paper trail for test performance
     PaperTrail.enabled = false
 
-    ::HmisUtil::JsonForms.seed_all if ENV['ENABLE_HMIS_API'] == 'true'
-
+    # disable rack attack unless we are testing it explicitly
     Rack::Attack.enabled = false
+
+    example_file_paths = RSpec.world.filtered_examples.values.flatten.map { |e| e.metadata[:file_path] }.uniq
+    # load fixpoints if we're running tests that may need them
+    if example_file_paths.grep(%r{/drivers/(hud_path_report|hud_spm_report|hud_data_quality_report)/}).any?
+      Dir.glob('{drivers,spec}/**/fixpoints/*.yml').each do |filename|
+        FileUtils.rm(filename)
+      end
+      Dir.glob('{drivers,spec}/**/fixpoints/*.sql').each do |filename|
+        # The 2024 test kit is HUGE, so we'll use the preprocessed version
+        # This means when the test kit changes, we need to update both the CSVs and the SQL fixpoint
+        next if filename.include?('drivers/datalab_testkit/spec/fixpoints')
+
+        FileUtils.rm(filename)
+      end
+      GrdaWarehouse::WarehouseReports::ReportDefinition.maintain_report_definitions
+    end
+
+    # load hmis forms if we're testing the driver
+    if example_file_paths.grep(%r{/drivers/hmis/}).any?
+      ::HmisUtil::JsonForms.seed_all if ENV['ENABLE_HMIS_API'] == 'true'
+    end
+
+    # GrdaWarehouse::Utility.clear!
+    # Delayed::Job.delete_all
+    AccessGroup.maintain_system_groups
   end
 end