[//]: # 'remove this if PR is for a release-* branch'
# _Please squash merge this PR_

## Description

The goal is to reduce the time required to run an individual test. On my local env, this was around 25 seconds. With these changes, the time goes to around 3 seconds.

changes:
* enable spring and update binstubs. Shaves off almost all the startup time for running a test or opening the console
* add logic to selectively load fixtures/form defs. This shaves off considerable time when running an individual test
* explicitly disable spring in CI since it just adds overhead there
* add spring-watcher-listen (fs events rather than polling to detect change). This works for me but maybe not for everyone?

example:
```ruby
#untracked/simple_spec_2.rb 
require 'rails_helper'

RSpec.describe 'hello world test', type: :model do
  it "is okay" do
    puts 'testing here'
    expect(1).to eq(1)
  end
end
```
```
/app $ time rspec untracked/simple_spec_2.rb 
Running via Spring preloader in process 30502
testing here
.

Finished in 1.82 seconds (files took 0.45802 seconds to load)
1 example, 0 failures

real    0m 2.75s
user    0m 0.22s
sys     0m 0.03s
```


## Type of change
[//]: # 'remove options that are not relevant'
* developer efficiency 

## Checklist before requesting review
- [x] I have performed a self-review of my code
- [x] I have run the code that is being changed under ideal conditions, and it doesn't fail
- [x] My code includes comments and/or descriptive variable names to help other engineers understand the intent (or not applicable)
- [x] My code follows the style guidelines of this project (rubocop)
- [x] I have updated the documentation (or not applicable)
- [x] If it's not obvious how to test this change, I have provided testing instructions in this PR or the related issue