@ttoomey 2024-09-30
> Do you have a sense of how much time overall this shaves off? It looks like this is most useful when running a handful of tests locally, which is also a big win.

It shouldn't impact overall CI run time. It's really just to allow rapid iteration on a single test. Once spring is started, the time to run one test is about 3 seconds (in my env).