@martha 2024-10-02
I rely pretty heavily on the Rubymine graphical debugger for running rspec tests, and this has broken something for me, but after spending about 1.5hr trying to debug it, I think I'm in over my head. I'll leave it for now and try to come back to it later. :(